+++ rusty_lr_core/src/parser/nondeterministic/context.rs
@@ -554,6 +554,158 @@ impl<Data: TokenData> Context<Data> {
         ret
     }
 
+    fn backtrace_impl<P: Parser<Term = Data::Term, NonTerm = Data::NonTerm>>(
+        &self,
+        parser: &P,
+        node: usize,
+    ) -> crate::Backtrace<&'static str, Data::NonTerm>
+    where
+        Data::Term: Clone,
+        Data::NonTerm: std::hash::Hash + Eq + Clone,
+    {
+        use crate::hash::HashSet;
+        use crate::rule::ShiftedRule;
+        use crate::rule::ShiftedRuleRef;
+        use crate::Backtrace;
+        use crate::Token;
+        use std::collections::BTreeSet;
+
+        // root
+        if self.nodes_pool[node].len() == 0 && self.nodes_pool[node].parent.is_none() {
+            let state0 = &parser.get_states()[0];
+            let mut rules = Vec::with_capacity(state0.get_rules().len());
+            for rule in state0.get_rules().iter() {
+                rules.push(ShiftedRule {
+                    rule: parser.get_rules()[rule.rule].clone(),
+                    shifted: rule.shifted,
+                });
+            }
+
+            return Backtrace {
+                traces: vec![rules],
+            };
+        }
+
+        let mut traces = Vec::new();
+        let mut current_rules: BTreeSet<_> = parser.get_states()[self.state(node)]
+            .get_rules()
+            .iter()
+            // .filter(|rule| rule.shifted > 0)
+            .copied()
+            .collect();
+        let mut next_rules = BTreeSet::new();
+        traces.push(current_rules.clone());
+        let mut zero_shifted_rules: HashSet<Data::NonTerm> = Default::default();
+
+        for state_idx in self.state_iter(node).skip(1) {
+            zero_shifted_rules.clear();
+            next_rules.clear();
+            for rule in current_rules.iter() {
+                if rule.shifted > 0 {
+                    next_rules.insert(ShiftedRuleRef {
+                        rule: rule.rule,
+                        shifted: rule.shifted - 1,
+                    });
+                    if rule.shifted == 1 {
+                        zero_shifted_rules.insert(parser.get_rules()[rule.rule].name.clone());
+                    }
+                }
+            }
+            std::mem::swap(&mut current_rules, &mut next_rules);
+            if zero_shifted_rules.is_empty() {
+                continue;
+            }
+
+            loop {
+                let len0 = current_rules.len();
+                for rule in parser.get_states()[state_idx].get_rules().iter() {
+                    let prod_rule = &parser.get_rules()[rule.rule];
+                    if let Some(Token::NonTerm(nonterm)) = prod_rule.rule.get(rule.shifted) {
+                        if zero_shifted_rules.contains(nonterm) {
+                            current_rules.insert(*rule);
+                            if rule.shifted == 0 {
+                                zero_shifted_rules.insert(prod_rule.name.clone());
+                            }
+                        }
+                    }
+                }
+                if len0 == current_rules.len() {
+                    break;
+                }
+            }
+            traces.push(current_rules.clone());
+        }
+
+        // node.iter() does not include root node.
+        // so explicitly add root node.
+        {
+            let state_idx = 0;
+            zero_shifted_rules.clear();
+            next_rules.clear();
+            for rule in current_rules.iter() {
+                if rule.shifted > 0 {
+                    next_rules.insert(ShiftedRuleRef {
+                        rule: rule.rule,
+                        shifted: rule.shifted - 1,
+                    });
+                    if rule.shifted == 1 {
+                        zero_shifted_rules.insert(parser.get_rules()[rule.rule].name.clone());
+                    }
+                }
+            }
+            std::mem::swap(&mut current_rules, &mut next_rules);
+            if !zero_shifted_rules.is_empty() {
+                loop {
+                    let len0 = current_rules.len();
+                    for rule in parser.get_states()[state_idx].get_rules().iter() {
+                        let prod_rule = &parser.get_rules()[rule.rule];
+                        if let Some(Token::NonTerm(nonterm)) = prod_rule.rule.get(rule.shifted) {
+                            if zero_shifted_rules.contains(nonterm) {
+                                current_rules.insert(*rule);
+                                if rule.shifted == 0 {
+                                    zero_shifted_rules.insert(prod_rule.name.clone());
+                                }
+                            }
+                        }
+                    }
+                    if len0 == current_rules.len() {
+                        break;
+                    }
+                }
+                traces.push(current_rules.clone());
+            }
+        }
+
+        Backtrace {
+            traces: traces
+                .into_iter()
+                .map(|rules| {
+                    rules
+                        .into_iter()
+                        .map(|rule| ShiftedRule {
+                            rule: parser.get_rules()[rule.rule].clone(),
+                            shifted: rule.shifted,
+                        })
+                        .collect()
+                })
+                .collect(),
+        }
+    }
+
+    /// Get backtrace infos for all paths.
+    pub fn backtraces<'a, P: Parser<Term = Data::Term, NonTerm = Data::NonTerm>>(
+        &'a self,
+        parser: &'a P,
+    ) -> impl Iterator<Item = crate::Backtrace<&'static str, P::NonTerm>> + 'a
+    where
+        Data::Term: Clone,
+        Data::NonTerm: Clone + Hash + Eq,
+    {
+        self.current_nodes
+            .iter()
+            .map(|&node| self.backtrace_impl(parser, node))
+    }
+
     /// Simulate parser and get next expected (terminals, non-terminals) for current context.
     /*
     pub fn expected_token<P: Parser<Term = Data::Term, NonTerm = Data::NonTerm>>(
@@ -596,17 +748,6 @@ impl<Data: TokenData> Context<Data> {
     }
 
 
-    /// Get backtrace infos for all paths.
-    pub fn backtraces<'a, P: Parser<Term = Data::Term, NonTerm = Data::NonTerm>>(
-        &'a self,
-        parser: &'a P,
-    ) -> impl Iterator<Item = crate::Backtrace<&'static str, P::NonTerm>> + 'a
-    where
-        Data::Term: Clone,
-        Data::NonTerm: Clone + Hash + Eq,
-    {
-        self.nodes().map(|node| node.backtrace(parser))
-    }
 
     /// move all nodes in `other` to `self`.
     pub fn append(&mut self, other: &mut Self) {
